requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==41.0.7
pytest==7.4.3
httpx==0.25.2
alembic==1.12.1 